+++ src/resource-metadata-sqs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## resource-metadata-sqs
 
 ### 0.2.0 / 02.04.2025
-* [Feature] CDS-1996 Add support for multiple regions
+* [Feature] CDS-1996 Add support for multiple regions and accounts
 
 ### 0.1.4 / 19.02.2025
 * [Fix] CDS-1876 Reduce EC2 batch size by half as default (50 --> 25) and let the user configure the chunk size